feat: Update plot scripts with extended x-axis range
The code changes in `plot_Bscan.py` and `plot_Bscan_CMP.py` modify the x-axis range of the plots. The extent of the images is multiplied by a factor of 4 instead of 2, resulting in an extended x-axis range. Additionally, the y-axis label in `plot_Bscan.py` is changed to "Time [ns]".

Based on the recent user commits and repository commits, it seems that these changes are related to improving the visualization of the B-scan plots.